Slightly restructure menu system to better allow navigating submenus.
Because menus can change via button callbacks instead of purely via the
handle_input() function, menus can now emit the menu_changed signal,
which directly calls the _change_menu() function without having to
somehow propagate that input to handle_input().

Additionally, UNPAUSED is now a proper state in the state machine,
greatly simplifying the logic surrounding menu navigation vs
pausing/unpausing. The UNPAUSED script is built into the dummy
UnpausedState VBoxContainer node, with the idea that it's special and
likely won't need to be modified much, but this might change in the
future.

Finally, as a proof-of-concept, the Quit menu was implemented with these
new changes to show that it's working, though a later commit will have
to address the issue of applying focus to the correct button when
navigating back to a menu (e.g. when selecting No from the Quit menu,
the Resume button on the Pause menu gets focus, instead of the Quit
button).