Implement basic controller remapping.
There are a couple of edge cases to iron out, such as handling joystick
inputs and removing previous mappings for other actions if they match
the new one. Also, controls should be saved and loaded from disk so that
the player doesn't have to reconfigure them every time they start the
game.

For completeness, the 'player_interact' input action was added, though
it currently does not do anything.