Add jump height multiplier to double jump.
Instead of being an identical second jump, the double jump now has only
70% of the initial jump's height. An important note is that the gravity
and jump duration are the same; only the jump height (and consequently
the max jump velocity) was modified. This ensures that the gravity
remains the same and that the player falls at the same speed regardless
of which jump they're falling from.