WIP wall slide.
Currently, wall detection is provided by a special node that
encapsulates a pair of raycasts. Before continuing it would probably be
a good idea to set up collision layers/masks on the player and
environment to simplify the hit detection, as well as modifying the
player's collision shape to more accurately reflect its body shape.