Only allow grappling to on-screen grapple points.
Note that this sometimes doesn't work; as far as I can tell there's a
sort of grid granularity that VisibilityNotifier2D uses, and it might be
large enough to make a noticeable impact in this low-res game. This
leads to scenarios where the grapple point is a tile or two off screen
but still ends up being grappleable.