Create GrappleType enum.
This commit introduces some basic skeleton code for adding a grapple
type to grapple points. The NORMAL grapple type causes grapple points to
behave as normal (i.e. the player is propelled to the grapple point
itself). The LAUNCH grapple type, when implemented, will instead cause
the player to grapple to points other than the position of the grapple
point itself. This is useful for e.g. crossing large chasms via a
grapple point positioned above the center of the drop.