Add skeleton code for 'hazard checkpoints'.
Rather than simply taking a normal hit, contact with hazards (e.g.
spikes) will now lead to a screen transition where the player's position
is reset to the last 'hazard checkpoint'. Currently, this checkpoint is
hard-coded, but eventually this will be decided programmatically based
on the player's progress towards making it to one end of the room.