Yield on additional time before letting player move elevator again.
Because the tween is moving the follow point, and we constantly
interpolate the position toward the follow point, the elevator will
still be moving for a bit when the tween completes. Waiting an
additional second before allowing the player to move the elevator again
ensures that the elevator will come to a complete stop before moving
again.