Add separate state for pausing during bash telegraph.
The idea is to split the telegraph into two distinct states: one for
actually telegraphing the attack via shaking, and one for pausing before
initiating the attack. This is useful for having the sentry drone track
the player's position during the shake, but then stop tracking during
the pause, giving the player a chance to escape the incoming bash.