Add "wobble" to laser beam.
Both the telegraph and shooting phases now incorporate "wobbling" in the
width of the inner/outer beams. This both looks nicer and allows for an
opportunity for the player to learn patterns and better predict when the
beam will fire, since the easing is non-linear for both sets of wobbles.
This is especially important since there is not yet any sound to provide
additional telegraphing.