Consolidate textures in tilesets/scenes/Platforms.tscn.
For some reason, some of the sprites in Platforms.tscn used a
StreamTexture that was semantically equivalent to the industrial.v2.png
texture used elsewhere, but was considered a separate resource. This
commit ensures that the resulting Platforms.tres tileset only uses the
one texture, but doing so resulted in losing all navdata. A future
commit will re-introduce this now-missing navdata.