Create GroundedPhysicsManager and use throughout.
This new physics manager is specific for all non-airborne actors
(player, sluggish failure, leaping failure, etc.). This allows various
calculations to be shared, such as for jump velocities and gravity. The
defaults are the player values so that entities that want to use those
(such as gravity, which must be calculated) can easily do so.