Clamp flash color values in flash shader.
Now that we have glow enabled for color values whose components exceed
1, it's possible for some flashes to inadvertantly trigger the glow.
This clamp ensures that the mix of the flash color and the original
sprite color never exceeds the glow threshold.

One alternate solution to this would be to avoid mixing the two colors
entirely and simply switch between original sprite color and flash
color.