Add basic FLY_TO_POINT state to ranged sentry drone.
Unlike the worker drone, this version uses actual nav data to fly to a
random point in the room, rather than sampling several points to find
one that has a clear line of sight path. This has the benefit of not
needing an obstacle detector to find valid paths, though future commits
will have to address the issue of the randomly-sampled point being
inside level geometry.