Change ranged sentry drone homing behavior.
Rather than perfectly following the player on every frame, have the
ranged sentry drone home in only slightly during the FOLLOW_PLAYER
state. This uses a technique similar to the homing projectile, except
that the homing weight is fixed in this case, as we always want the
drone to be following the player in some way during the FOLLOW_PLAYER
state.