Use separate sprites for inner/outer laser beams.
Using separate sprites for the inner and outer beams allows for proper
blending when trying to "antialias" the inner beam.

Previously, increasing the transparency of the pixels of the outer edge
of the inner beam (via smoothstep()) caused those pixels to blend with
the background, instead of with the outer beam. This is because both
beams shared a sprite/shader, so the background was always underneath.

Now, the inner beam is a separate sprite that sits on top of the outer
beam sprite. This causes the pixels of the outer edge of the inner beam
to correctly blend with the outer beam sprite underneath.