Create basic state sequence.
The idea behind the state sequence is that it acts as a state that
simply delegates to each of its children states in sequence. If the
current active child state returns a non-NO_CHANGE state, the state
sequence proceeds to the next state in the sequence. Once the last state
in the sequence returns a non-NO_CHANGE state, that state is returned
from the state sequence and acts as the state to transition to from the
overall state sequence.

At each stage, the additional metadata that was initially passed to the
state sequence is propagated to each state in the sequence, which means
that all additional metadata that any of the states in the sequence may
need must be passed to the state sequence as part of the initial
change_state() call.