Fix state sequence bug.
Need to reset active state to the first child every time the sequence is
entered, otherwise sequences will only run the last state's code if the
sequence is repeated (e.g. reading a sign for the second time skips the
WalkToPoint state and only runs the code for the FaceObject state).