Move Menu enum to ui/menus/menu.gd.
This is in preparation for adding a title screen with a title menu, the
hope being that we can reuse the existing options/quit menus from within
the title menu as well. It also make more sense to have menu.gd own the
menu enum, especially since we will soon have non-pause menus.