Use pushdown automaton for navigating menus.
This is something that probably should have been used from the
beginning, but it wasn't strictly necessary before because there was
only one set of menus (i.e. the menus reachable by pausing the game).
Now that we are going to include another set of menus from the title
screen, it would be nice to share some of the ones that should be the
same for both the title screen and the pause screen (e.g. options menu).

This wasn't possible previously because the transitions back to old
menus were hardcoded (since there was only ever one way to go back
before). This commit introduces a new special PREVIOUS menu enum that is
now used to indicate that the user wants to go back to the previous
menu, without having each submenu know which menu the previous one is.
Advancing forward to submenus also no longer requires that each menu
pass the current menu (i.e. itself) in the change_menu() call, which has
been renamed to advance_to_menu() for further clarity.