Implement very basic title screen system.
The title screen behaves similarly to the pause menu, except that it
also has an instance of Game.tscn that it switches to when the player
presses the start button. This is an exported variable and thus
configurable, so multiple such scenes could be created that each point
to a different Game scene, while still retaining the ability to run Game
scenes directly for debugging purposes (in order to skip the title
screen).