Fix onready type hints.
Seems as though Godot 3.2.2 now correctly identifies mismatched types in
onready var type hints when running the main game scene, which wasn't
happening in 3.2.1. This commit fixes the type hints to match the actual
node types in question, though I wonder how this was ever working in the
first place.