Fix HSlider highlight bug.
It seems as though Godot 3.2.2 introduced a new "grabber area highlight"
class item style for HSlider. Since this style obviously didn't exist
beforehand, highlighting the audio sliders in the audio options menu
caused the default style to kick in, which clashed with the custom
theme. This commit introduces an empty style for grabber area highlight
to prevent this.