Slightly improve HSlider theme styles.
Introduce a separate highlight color (same as button hover color) and
decrease the width of the slider style box line.

The width of the slider style box line seems to be different on the left
and right side of the grabber icon, would like to fix this but I can't
seem to figure out how for now.