Fix infinite WALK_TO_POINT loop.
The WALK_TO_POINT state previously would infinitely loop if the player's
current position was the stopping point. This can manifest when e.g. the
player attempts to rest immediately after loading the game (so the
player's current position is already the rest position used as the
stopping point in the REST_AT_LAMP sequence).