Fix rest-at-lamp bug.
Ensure that all input-related actions in lamp's _unhandled_input() call
are guarded by a check that the lamp in question is the one near the
player. This prevents one lamp from performing actions that override the
lamp near the player, since only the lamp near the player has its input
disabled while resting.