Ensure active state in sequence always calls exit.
Previously, if a state sequence was interrupted by having the player
state changed via elsewhere in the code, the current active state in the
sequence would not have its exit function called. This happened when
using player_interact while resting to go back to idle; this would
cause the z's sprite to appear anyway, since REST's exit function was
not called, which would normally hide the sprite.