Only save options at fixed points.
Specifically, when going back to the previous menu from an options menu
or when resetting options to their defaults.

This is mainly for addressing a bug when running the game on Windows
where moving the audio sliders (which saved options every time they
changed) too quickly would cause the game to crash with an error when
trying to save to disk. My current working theory is that it tried to
open the file while it was already open.