Add LOOK_DOWN and LOOK_UP player states.
These states pan the camera up or down to allow for the player to better
plan complex platforming with platforms out of view.

Currently there's an issue with keyboard bindings, as the up-arrow key
is also used for up attacks (as opposed to controllers that use a
different analog stick). Will probably need to add a timer for keyboard
users so that they have a chance to up-attack before panning the camera
upwards.