Add basic player death transition code.
This commit introduces two major changes to create a player death
transition effect:

1) A shader responsible for the actual screen transition effect that
   simply expands a black circle from the player center out to the edges
   of the screen

2) An improved on-death callback that initiates the death transition
   effect, resets the world, and other small things like rumble and
   screenshake

This code is incomplete: it somewhat-jankily reuses the exact callback
for resting at a lamp to reset the world, which means that the game will
crash if the player dies before reaching a lamp. The death animation for
the player itself is also quite basic and will be improved upon in
subsequent commits.