Change StreamTexture to Texture in Platforms.tres.
For whatever reason, the underlying texture in the platforms tileset was
a StreamTexture instead of a regular Texture, which apparently doesn't
get exported by default like a regular Texture. This change allows the
platforms tileset to be correctly exported.

Note that this change was done manually (i.e. not via the editor): the
StreamTexture sub-resource in the .tres file was manually replaced with
a Texture external resource, with the individual tiles updated
accordingly. The editor then automatically updated the sub-resource ids,
as there is now one fewer sub-resource.