Fix assertions with side effects.
All assertions on expressions that have important side effects have been
separated into a function call and an assertion on the result. The
reason we do this is because assertions are removed in release builds,
so important side effects don't happen. This was discovered while
testing an export build where changing scenes to the title screen didn't
work because the underlying change_scene_to() call was wrapped in an
assert.