Set mouse capture mode when pausing/unpausing.
Currently, capture mode is set to CONFINED when paused (i.e. visible but
unable to leave the game window) and CAPTURED when unpaused (i.e. hidden
and unable to leave the game window).

In the future, it may be desirable to use VISIBLE instead of CONFINED
when paused (i.e. where the mouse can leave the game window), as this
might be slightly more convenient when playing in windowed mode.
However, this is also slightly more complicated, as we'd have to prevent
mouse events outside the game window when VISIBLE from affecting the
game, as well as re-capture the mouse input when clicking on the game
window.