Fix AbilitiesLamp room save bug.
Previously, the condition used to determine whether to keep the door in
the AbilitiesLamp room closed was whether the player had entered the
room. However, this caused a bug where the player could quit out after
triggering the door to close but before resting at the lamp and find
themselves unable to re-enter the room.

The new condition is therefore resting at the lamp, which will cause the
player to spawn in that room, ahead of the closed door.