Add skeleton code for ability-specific rooms.
This commit introduces the basic framework for generating rooms based on
which ability the player chooses. The idea is to maintain a list of
room-position pairs for each ability so that they can be added to the
tree at runtime once the player chooses an ability.

Note that Game._process() had to be modified to get the list of children
in the World/Rooms node anew each frame to avoid weird behavior
associated with changing the _rooms array while iterating over it.