Attempt fixing stupid door closing animation bug.
Because the original bug is difficult to consistently reproduce, it's
difficult to know whether this actually fixed the underlying issue.
However, even if it is eventually discovered not to have fixed it,
having separate animations for open and close is a bit cleaner than
playing one backwards and one forwards.