Use WALK as initial state for leaping failures.
Because enemies both reset and pause when the player exits a room, the
leaping failures would previously always have a dust puff show up at
their feet the moment the player entered the room, as they were
restarting in the old FALL default state. Using WALK avoids the dust
puff and still allows the failure to start in midair, as there is a
transition to the FALL state in that scenario.