Only show the player's current room.
All rooms now start off hidden, and only become visible once the player
enters. The room will then become hidden again once the player exits.

From what I've gathered, godot will attempt to cull/clip offscreen
objects so as not to draw them, but the act of culling/clipping is
itself work, work that isn't done on objects that are hidden. This
commit attempts to prevent doing unnecessary culling/clipping work by
hiding everything that's not in the player's current room.

When starting the demo from scratch, the number of 2D draw calls is
roughly the same, perhaps even slightly higher, with this new
optimization (for whatever reason). However, the number of draw calls is
reduced by half or more when starting in an ability-specific generated
room.