Improve room transition performance.
Previously, checking for a room transition required iterating over all
the rooms in the scene tree per frame and checking whether they
contained the player. Now, a single pixel-sized Area2D attacked to the
player serves as a "room detector", so each room can know when the
player has entered with a single area_entered signal callback. Because
pixel snapping is enabled, this single pixel-sized area cannot be in two
rooms at the same time.