Fix bug in new room transition system.
The room detector area on the player was not previously part of the
mirror_y_axis node group, meaning that its position would not be
mirrored along the y-axis when the player changed direction. This could
therefore cause issues with room transitions if the player immediately
tried to re-enter the previous room.

However, this fix alone was not enough to fix the bug, and the reason
why eludes me. Perhaps the one-pixel-sized area can in fact occupy more
than one room at a time, even with pixel snapping enabled? In any case,
an additional fix was added that simply moves the player one pixel into
the direction of the new room during the transition. This effect is not
noticeable in-game, and seems to fix the previous issues, so for now we
consider this bug fixed.