Create tutorial message skeleton.
The TutorialMessage node is a new top-level UI node that displays either
a control (i.e. explaining controls) or non-control message. The control
message includes a sort of "oneof" node that is either a texture for
controller buttons or a text box for keyboard buttons. The newly-added
Controls autoload will automatically detect when the input type changes
from keyboard to controller, and the associated signals can be used to
automatically update the TutorialMessage node to switch between the two
"oneof" states.

Future commits will introduce the special Area2Ds to trigger the
messages at specific points in the game.