Further streamline Controls API.
This commit primarily introduces a new get_joypad_texture_for_action()
function in the Controls autoload, which will directly return the joypad
texture for a specified player action. This is especially useful since
it will choose the correct texture depending on whether the action is a
joypad button or a joypad motion (i.e. left or right stick).