Add special cases for moving and looking up/down.
Because moving and looking up/down are one controller stick but two
keyboard buttons, include a special case for turning any of the
corresponding player actions into concatenated versions of the combined
controls for the keyboard control button label.

For example, for a tutorial message trigger area corresponding to the
player action 'player_move_left', the associated keyboard button label
in the tutorial message will be 'Left/Right' instead of just 'Left'.

It might be possible to accomplish this by instead having the trigger
areas export multiple player actions, but then the controller side of
things becomes more complicated, since that deals with textures instead
of strings.