Revert "Calculate grapple points once upon entering room."
This reverts commit dbdd50316f7f79292aad81a05663329a1b22ffce.

Turns out this "optimization" leads to a bug when the player starts out
in a room with grapple points, as the code for setting the grapple
points will not have run by the time the player spawns (since the room
transition code is responsible for setting the grapple points).

Reverting is simpler than finding a good spot to calculate grapple
points once at the start, since this "optimization" likely didn't have
any significant impact anyway.