Show all rooms all the time.
This effectively reverts the previous commit where only the player's
current room would be shown on the screen. This behavior, while
effective at reducing draw calls, caused visual artifacts whenever parts
of the previous room needed to be shown while the player was not
actually in them (e.g. screenshake).

A more sophisticated version of this scheme would also show the adjacent
rooms, but it's currently difficult to automatically determine adjacent
rooms in this current setup, so if this is still desireable later this
more robust approach may become worthwhile.

A more hacky short-term approach would be to change the default fill
color to be the background color, but this still wouldn't be ideal, as
platforms that continue into the next room would appear to be cutoff
during e.g. screenshakes.