Distinguish between fall and leap down.
For LeapingFailure, rename JUMP and FALL to LEAP_UP and LEAP_DOWN to
clarify that these are the two phases of the LEAP sequence. The single
FALL state now has a new script that behaves identically to LEAP_DOWN
except that it doesn't include motion in the x-axis, which makes it
better suited for falling off ledges.