Allow hazard checkpoints to work when ungrounded.
That is, allow hazard checkpoints to register the player's presence even
before the player touches the ground. This is useful for when we want to
ensure that a player will activate a hazard checkpoint (e.g. at the
start and end of a room with hazards), even if they e.g. dash through it
in the air. This would cause the current hazard checkpoint not to
activate and may result in a crash if the player hasn't activated any
hazard checkpoints prior to this (or they might respawn in a prior room
if they activated one before then).