Fix LeapingFailure spinning bug.
Remove the function call to face the player in each frame, as this led
to the failure constantly spinning in place in certain situations
(player standing directly above failure, failure hitting wall and
immediately turning around to face player, etc.).

The existing call to face the player in expand/contract's enter()
function suffices.