Create credits screen skeleton.
This commit introduces the simple skeleton for the credits screen. It
currently populates a ScrollContainer with several fake Labels and
autoscrolls them until it reaches the end, at which point it
automatically switches back to the title screen. The player can also
press any key/button during the autoscroll to go back to the title
screen early.