Add Version helper script for game versioning.
This script is a series of static functions instead of an autoload, as I
fear that I'd have to fiddle with the autoload ordering were I to make
it an autoload (since it will end up being used in existing autoloads
like SaveAndLoad and Options). All we need are a few static functions
anyway, since we're just reading from ProjectSettings.