Do not hover when attacking in midair.
This commit is the first in a series meant to tweak attacking to feel a
little more comfortable, especially in midair. The idea is to eventually
arrive at a point where attacks only "hover" when they actually connect
with an enemy. When attacks don't connect, the player can move freely
while airborne similar to falling. Note that this only applies to
airborne attacks; attacks while grounded do not allow the player to move
at all.

This commit specifically removes the "hover" effect while attacking
normally and therefore allows the player to continue to move while
attacking in midair.