Fix "skate" bug when hitting A and X together.
Since the entire velocity vector was being taken from the previous JUMP
state in the new ATTACK/ATTACK_UP code, this including the x-component,
which we specifically only want to allow while the player is airborne.
Hitting A and X at the same time would cause the player to enter the
JUMP state and then immediately enter one of the ATTACK states with a
non-zero x-component in the passed-in velocity.

The solution is simply to zero out the x-component in enter(), thereby
ensuring that the only way we get x-velocity while attacking is if
airborne as detected in update().