Allow jumping right after attack animation ends.
If the player presses jump during an attack animation and they're
otherwise allowed to jump, transition to JUMP after the attack animation
ends.

Note that it's required that the jump button continue to be held at the
end of the attack animation to allow the player to jump cut later.