Fix slight wall jump bug.
Previously, the player would not be flipped away from the wall if one of
the wall detector raycasts was not hitting the wall. This commit fixes
that bug by turning the player away from the wall unconditionally.

It may be worthwhile investigating later whether being near a wall
should necessarily mean that BOTH raycasts hit the wall (as is currently
the case), instead of just ONE of the raycasts (or both).