Use new abstractions for error handling.
This commit defines two new abstractions for handling errors:
ErrorPlusMessage and ErrorOr. The former is essentially an Error
constant with an accompanying error message, while the latter
encapsulates a value and an ErrorPlusMessage. The idea of ErrorOr is
that it will either have a value, or it will have an ErrorPlusMessage.

These abstractions replace the existing _handle_error() functions, as we
will need to handle errors at callsites instead of as soon as they
happen, thereby necessitating that we propagate errors all the way to
the top-level functions in Options and SaveAndLoad.