Display load error on save slot.
Privatize has_valid_version() and call SaveAndLoad.load_game() in each
save slot container for the corresponding save slot. This will allow us
to pass possible errors to the save slot error menu in the future.

We don't care about the fact that a long error message causes the save
slot menu to look weird, since we'll soon display a generic message on
the save slot itself and then allow the user to click the slot and be
taken to the dedicated save slot error menu.