Create SaveSlotErrorMenu.
Rather than trying to show the error on the slot text itself, allow the
player to click the slot anyway and transition to a new save slot error
menu that displays the full error message and error code. The slot is
tinted pink to further highlight that it's an invalid save file.