Replace OS cursor with custom "software" cursor.
Because custom cursors set via ProjectSettings do not respect the
in-game resolution, resort instead to a custom "software" cursor (i.e. a
sprite whose position tracks that of the actual cursor every frame).

We have to compromise a bit here, as MOUSE_MODE_CONFINED no longer works
as intended, since it will show the actual OS cursor. Therefore, we have
to use the equivalent of MOUSE_MODE_VISIBLE for the title screen
instead.

Note that the ACTUAL mode set in cases where we would conceptually use
MOUSE_MODE_VISIBLE is MOUSE_MODE_HIDDEN, as we want the actual OS cursor
to be hidden. The custom cursor will determine separately whether it
should be visible or not (by checking the actual mouse mode).