Create AudioStreamPlayerVisibility skeleton.
The idea behind this node is to tie an audio stream's volume to its
on-screen visibility: audio should play while the parent node is visible
and quickly attenuate to muted once the parent node leaves the screen
(and vice versa once it starts to come back). This will be achieved with
a pair of VisibilityNotifier2D nodes:

* ObjectVisibility will encompass the parent node
    * on screen entered, play audio at max specified volume
    * on screen exited, start attenuation processing (fade out)

* AttenuationVisibility will be larger than the parent node and will
  effectively control the attenuation distance
    * on screen entered, start attenuation processing (fade in)
    * on screen exited, mute audio

This corresponds for three states that the AudioStreamPlayerVisibility
node can be in:

* VISIBLE (audio plays at full volume)
* ATTENUATING (audio attenuates based on distance to screen edge while
  off screen)
* INVISIBLE (audio muted)

This is only important for rooms that are larger than a single viewport,
as they are the ones with the possibility of having e.g. an off-screen
enemy we only want to start hearing as we approach it. All audio
processing will stop once the player leaves the parent node's room.