Add volume logic to AudioStreamPlayerVisibility.
Because getting the viewport's global position is harder than I thought,
the attenuation is sort of faked by simply tweening the volume to/from
50% max_volume_db loudness while attenuating. This has a couple of
drawbacks but seems good enough in testing for now.

One thing to note is that this implementation naively tweens the
underlying audio stream player's volume_db property directly, rather
than tweening some other value linearly and calculating the nonlinear db
change at each tween step (like ScreenFadeout). In testing this seemed
to be mostly ok but future implementations might try a similar approach
as ScreenFadeout (or even possibly tweening nonlinearly using an
inverse-ish curve to the db curve).

Additionally, it may be possible to keep track of the velocity of the
parent node and use that to inform how quickly/slowly we fade the volume
in and out.