Add blank death animations to all enemies.
Now that we're adding audio (and some of that audio is tied to
animations), we need a blank death animation to play on death so that
non-death animations don't emit sound effects as they're stopped and
started on room enter/exit after death.

Also has the benefit of already being there if we decide to add actual
animations for death in the future.