Add compressor to effects bus.
The default version is already pretty good at mitigating the effect of
multiple identical sound effects (e.g. multiple failures walking in
tandem) having an increased volume.

It's possible it could be even better with some tweaking but for now
we'll keep it at the default state and re-assess once we add the rest of
the sound effects.