Expand Audio API and fix minor related bugs.
The new API allows for setting the max linear volume per-bus and also
resetting the volume of a bus to its max value. This makes it easier to
change the volume of a bus in some way (e.g. muting/fading) and then
resetting that bus volume to its previous max volume (as set in the
audio options menu).

This commit also fixes the related bug where unpausing the game reset
the music bus to full volume, instead of the volume specified by the
user in the audio options menu.