Attenuate AudioStreamPlayerVisibility correctly.
This replaces the previous hacky tween-based attenuation with an actual
per-frame attenuation implementation. The weighting function (used for
calculating linear volume while attenuating) is 1 when the object is
just offscreen and 0 once the attenuation rect is also offscreen. This
value is clamped just to be sure, since the calculation may be slightly
off due to the resolution of the visibility notifier cell size.