Add all four main sector entrances.
Sectors 1 and 2 will be accessible from the moment the player lands in
the central hub. Sectors 3 and 4 will require both abilities gained from
sectors 1 and 2 (dash and wall jump, respectively) to proceed past their
respective entrances.

While this design choice will somewhat limit player freedom in terms of
the order in which they visit each sector, it will hopefully enhance the
quality of the later two sectors, as each can now be designed with the
player having these two abilities in mind.