Use min-jump buffer jump instead of max-jump.
This fixes the slightly annoying phenomenom where the player activates
buffer jump but doesn't keep the jump button pressed when entering the
jump state, resulting in a max jump (since the button cannot be released
to trigger the jump cut).

In the future, we could consider lerping between min-jump and max-jump
with a small weight, should min-jump prove to be too low in practice.