Implement barebones springboard functionality.
For now, the player simply transitions into the JUMP state upon
activating the springboard. This comes with all the pros (reusing code)
and cons (min/max jump heights aren't really appropriate anymore) of
such an approach.

Improving on this may be as simple as allowing min/max jump heights to
be passed into the JUMP state, though if we want to add more
advanced/nuanced behavior (like momentum calculations), it may be better
to just create a new state for this. This will help prevent the JUMP
state having to itself keep state about whether it's a normal jump or a
spring jump.