Fix crush detection logic.
For some reason, after upgrading to Godot 3.3, the existing crush
detection logic would not work in the Demo game instance (though it
continued to mostly work in the CrusherGauntletDemo instance for some
reason). Swapping the logic (and using Area2D instead of RayCast2D for
proximity checks) seems to have fixed it with no discernable downsides
(note that the extents of the area are narrower to prevent the previous
issue with walking into the side of platforms triggering crushes).