Create SectorOne_17.
This is where the player would acquire the dash ability, which I'll add
later.

The current setup of this room also technically allows for using the
wall jump to bypass getting the dash, though subsequent rooms will
DEFINITELY require the dash so I don't really see this as a problem.