Switch to 'live counting' for arena enemies.
Rather than maintaining a separate variable updated in enemy death
callbacks, simply count the number of live non-projectile enemies every
frame to determine when to move to the next phase.

This is meant to mitigate the bug where under SOME set of unknown
circumstances, getting hit by a laser in the grapple arena would cause
the arena to be stuck on phase two (i.e. all the enemies would die but
the next wave wouldn't respawn and the door wouldn't open). While the
exact circumstances are unknown, it is known that this is due to the
previous separate variable not being updated during an enemy death, and
therefore stuck at 1 at the end of the phase instead of 0.