Create basic hazard water skeleton.
The idea behind the hazard water is that we can place one in a scene and
then simply position/resize the collision shape to represent the body of
water. All the visual effects will then automatically reposition/resize
themselves to conform.

This first commit includes a sprite for the body of water and a simple
gradient shader. Future commits will include things like surface foam,
bubble particles, etc.