Improve camera for room transition corner cases.
The new camera will interpolate from the camera's current global
position to the "best" position in the next room, rather than going
between camera anchors. "Best" here means it will attempt to slide in
the appropriate direction (based on the direction of the transition),
while respecting the new room's camera limits.

This behavior is idential to the old behavior for the majority of
scenarios, but will correctly (i.e. smoothly) transition in the handful
of cases where the camera would be centered somewhere other than the new
room's anchor (MainLift lower left if jumping, SectorTwo_1, etc.).

Note that the anchors are still needed in order to guide the camera in
the event that it would otherwise land outside of the new room's bounds.