Fix camera pan + transition bug.
Now that we're using the camera's position as the start of the
transition tween, it's susceptible to incorrectly using the mid-pan
position if we transition mid-pan. Therefore, ensure that the transition
tween is reset and the position reset to zero before transitions. This
also required replacing the use of get_camera_position() with a simple
global_position for the starting camera position, as that function
didn't seem to take the newly-reset position into account.