Create audio group node for crushers.
Similar to the conveyor belt audio group, this node helps prevent
overlapping sounds from becoming too loud too quickly. In this case
though, we need separate audio group nodes for every combination of
speed and delay for all crushers in that room, so it's still possible
that SOME sounds might overlap, but it's a lot better than before.